+++ pytsite/auth/__init__.py
@@ -6,7 +6,7 @@ from ._api import get_current_user, get_user_statuses, get_user, create_user, ge
     register_auth_driver, user_nickname_rule, sign_in, get_auth_driver, create_role, get_sign_in_url, get_sign_out_url,\
     verify_password, hash_password, sign_out, get_access_token_info, switch_user, get_anonymous_user, \
     get_system_user, get_users, get_storage_driver, count_users, count_roles, get_first_admin_user, get_roles, \
-    get_user_modify_form, base_path, switch_user_to_system, switch_user_to_anonymous, restore_user
+    get_user_modify_form, base_path, switch_user_to_system, switch_user_to_anonymous, restore_user, prolong_access_token
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -60,7 +60,7 @@ def __init():
 
     # Event handlers
     events.listen('pytsite.setup', _eh.setup)
-    events.listen('pytsite.router.dispatch', _eh.router_dispatch, priority=-9999)
+    events.listen('pytsite.router.dispatch', _eh.router_dispatch, priority=-999)
     events.listen('pytsite.router.response', _eh.router_response)
 
     # Console commands