+++ pytsite/router/_api.py
@@ -336,25 +336,21 @@ def dispatch(env: dict, start_response: callable):
 
     except Exception as e:
         if isinstance(e, _HTTPException):
-            # If exception has embedded response in its body
+            # Exception can contain response object in its body
             if isinstance(e.response, _http.response.Response):
                 # For non-redirect embedded responses use original status code from exception
                 if not isinstance(e.response, _http.response.Redirect):
                     e.response.status_code = e.code
-
                 return e.response(env, start_response)
-
-            code = e.code
-            title = _lang.t('pytsite.router@http_error_' + str(e.code))
-            _logger.error('HTTP {} {} ({}): {}'.
-                          format(e.code, e.name, current_path(resolve_alias=False, strip_lang=False), e.description))
+            else:
+                code = e.code
+                title = _lang.t('pytsite.router@http_error_' + str(e.code))
+                _logger.error('HTTP {} {} ({}): {}'.format(
+                    e.code, e.name, current_path(resolve_alias=False, strip_lang=False), e.description))
         else:
             code = 500
             title = _lang.t('pytsite.router@error', {'code': '500'})
             _logger.error(e, exc_info=e)
-
-        from pytsite import metatag
-        metatag.t_set('title', title)
 
         args = {
             'title': title,